Further improved whitespace
Whitespace is still not part of the AST during analysis (simplifies
things a fair bit), but now it is stored under components. This allows
the output to better match the input.

One issue: whitespace after the first element of a set of Parens is
still hardcoded. I could fix this by removing the 'name' attribute of
Parens.